requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
apscheduler
pytest
httpx
cryptography